feat(forward): add max connect attempts knob
In #7704 we fixed a busy loop error by coupling upstream connection
errors with a shared counter. This counter collides with healthchecks
and the max_fails parameter, and it might have unintended side
effects.

This change replaces that implementation with an optional per-request
counter. It is limited by the new max_connect_attempts parameter.
By default this is set to 0, maintaining the legacy behaviour and
thus keeping it backwards-compatible. Updates tests and plugin docs.

Signed-off-by: Ville Vesilehto <ville@vesilehto.fi>